% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gasmet_data.R
\name{check_gasmet_columns}
\alias{check_gasmet_columns}
\title{Check column existence in Gasmet TXT files}
\usage{
check_gasmet_columns(
  gasmet_data_list,
  necessary_cols = c("Date", "Time", "Carbon.dioxide.CO2"),
  extra_cols = c("Nitrous.oxide.N2O", "Methane.CH4", "Ammonia.NH3", "Carbon.monoxide.CO",
    "Water.vapor.H2O"),
  ...
)
}
\arguments{
\item{gasmet_data_list}{List of data frames from Gasmet TXT files, read using
\link{read_gasmet_files}.}

\item{necessary_cols}{Character vector of column names which need to be
present in each Gasmet data frame for data processing to proceed. Names of
files not containing these columns will be compiled into the returned
"fatal" message.}

\item{extra_cols}{Character vector of column names which do not need to be
present in every Gasmet data frame for data processing to proceed, but
which should be checked anyway. Names of files not containing these columns
will be compiled into the returned "nonfatal" message.}

\item{...}{Ignored.}
}
\value{
A list of character vectors with names 'fatal' and 'nonfatal'.
}
\description{
Check for necessary and other expected columns in data from Gasmet TXT files.
Compile messages listing files that do not contain specified columns.
}
