% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gasmet_data.R
\name{convert_dates_and_times}
\alias{convert_dates_and_times}
\title{Convert date and time columns to specified formats}
\usage{
convert_dates_and_times(
  data,
  date_col = NULL,
  date_formats,
  time_col = NULL,
  time_formats
)
}
\arguments{
\item{data}{A data frame}

\item{date_col}{Name of date column in \code{data}. Set to NULL if no date column
exists.}

\item{date_formats}{Possible formats of dates in \code{date_col}. Passed to
\code{\link[lubridate:as_date]{lubridate::as_date()}}.}

\item{time_col}{Name of time column in \code{data}. Set to NULL if no time column
exists.}

\item{time_formats}{Possible formats of times in \code{time_col}. Passed to
\code{\link[readr:parse_datetime]{readr::parse_time()}}.}
}
\value{
A data frame
}
\description{
Convert date and time columns to specified formats
}
