% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_NA.R
\name{count_NA}
\alias{count_NA}
\title{Print a message with count and percent NAs in a numeric vector}
\usage{
count_NA(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Nothing
}
\description{
Print a message with count and percent NAs in a numeric vector
}
\examples{
count_NA(c(1, NA, 2, 3, 4, NA))
}
