% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gasmet_data.R
\name{find_gasmet_files}
\alias{find_gasmet_files}
\title{Find Gasmet TXT files}
\usage{
find_gasmet_files(
  dir = getOption("gaseous.gasmet_txt_dir"),
  recursive = TRUE,
  rm_open_files = TRUE,
  ...
)
}
\value{
A character vector.
}
\description{
List all Gasmet TXT files in specified directory. By default, the searched
directory is \verb{data/00_raw/gas_concentration}.
}
\examples{
if (requireNamespace("gaseous")) {
  ex_gasmet_dir <- system.file("example_data/data/00_raw/gas_concentration", package = "gaseous")
  find_gasmet_files(ex_gasmet_dir)
}
}
