% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gasmet_data.R
\name{import_gasmet_data}
\alias{import_gasmet_data}
\title{Import data from Gasmet TXT files}
\usage{
import_gasmet_data(
  dir = getOption("gaseous.gasmet_txt_dir"),
  recursive = TRUE,
  rm_open_files = TRUE,
  col_types = NULL,
  necessary_cols = c("Date", "Time", "Carbon.dioxide.CO2"),
  extra_cols = c("Nitrous.oxide.N2O", "Methane.CH4", "Ammonia.NH3", "Carbon.monoxide.CO",
    "Water.vapor.H2O"),
  file_date_col = "Date",
  file_date_formats = c("\%m/\%d/\%Y", "\%Y-\%m-\%d"),
  file_time_col = "Time",
  file_time_formats = c("\%H:\%M:\%S"),
  path_n_date = -2,
  path_date_formats = c("\%Y\%m\%d"),
  return_date_tests = FALSE,
  ...
)
}
\arguments{
\item{dir}{Directory to search for TXT files. By default, search
path is "data/00_raw/gas_concentration". This value can be set globally
using the option "gaseous.gasmet_txt_dir".}

\item{recursive}{Logical. Should \code{dir} be searched recursively? Defaults to
\code{TRUE}.}

\item{rm_open_files}{Logical. Should open files be removed from the list?
Defaults to \code{TRUE}.}

\item{col_types}{List or col_spec object compatible with col_types argument
of \code{\link[readr:read_delim]{readr::read_tsv()}}, identifying names and types of columns to be read
in from Gasmet TXT files. If set to \code{NULL}, a default column specification
will be used which includes Date, Time, and all gas ppm columns.}

\item{necessary_cols}{Character vector of column names which need to be
present in each Gasmet data frame for data processing to proceed. If all of
these columns are not found in every Gasmet TXT file, an error will be
thrown with information about which files are missing which columns.}

\item{extra_cols}{Character vector of column names which do not need to be
present in every Gasmet data frame for data processing to proceed, but
which should be checked anyway. Names of files not containing these columns
will be displayed in a message.}

\item{file_date_col}{Name of date column in Gasmet TXT files. Defaults to
\code{"Date"}, which should not be changed in most cases.}

\item{file_date_formats}{Character vector of possible date formats to parse
within TXT file data. Date parsing will be attempted in the order of
formats specified using \code{\link[lubridate:as_date]{lubridate::as_date()}}. See \code{\link[=strptime]{strptime()}} for
formatting abbreviations.}

\item{file_time_col}{Name of time column in Gasmet TXT files. Defaults to
\code{"Time"}, which should not be changed in most cases.}

\item{file_time_formats}{String of time format to parse within TXT file data.
Date parsing will be attempted using \code{\link[readr:parse_datetime]{readr::parse_time()}}. See
\code{\link[=strptime]{strptime()}} for formatting abbreviations. Defaults to \verb{\%H:\%M:\%S}.}

\item{path_n_date}{Integer. Location of date in path to Gasmet TXT file when
split by the file separator. Defaults to -2, which corresponds to the
directory containing the TXT file. Passed to \code{\link[stringr:str_split]{stringr::str_split_i()}}.}

\item{path_date_formats}{Character vector of possible formats to parse dates
extracted from file paths. Date parsing will be attempted in the order of
formats specified. See \link{strptime} for formatting abbreviations.}

\item{return_date_tests}{Logical. Should the date column test results be
returned instead of Gasmet data? Used for debugging only.}

\item{...}{Ignored.}
}
\value{
If \code{return_date_tests} is FALSE (the default), returns a list of
tibbles (one per gasmet TXT file). Otherwise, returns a data frame of date
test results.
}
\description{
Import and perform preliminary checks and validations on Gasmet TXT data.
}
