% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outlier.R
\name{is_outlier}
\alias{is_outlier}
\title{Detect outliers}
\usage{
is_outlier(x, n_IQR = 3)
}
\arguments{
\item{x}{A numeric vector in which to detect outliers}

\item{n_IQR}{Integer. Number of interquartile ranges below 25\% quantile or
above 75\% quantile at which a value will be considered an outlier. Defaults
to 3.}
}
\value{
A logical vector the same length as \code{x}
}
\description{
Detects outliers in a vector as values outside a specified
number of interquartile ranges below the 25\% quantile or above the 75\%
quantile.
}
