% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_collar_height.R
\name{mean_collar_height}
\alias{mean_collar_height}
\title{Calculate grand mean collar height from dates of collar changes only}
\usage{
mean_collar_height(
  data,
  collar_var = collar_height_cm,
  SU_var = SU,
  date_var = Date
)
}
\arguments{
\item{data}{A data frame}

\item{collar_var}{<\code{\link[rlang:args_data_masking]{data-masked}}> Name of collar
height variable in \code{data}}

\item{SU_var}{<\code{\link[rlang:args_data_masking]{data-masked}}> Name of sampling
unit variable in \code{data}}

\item{date_var}{<\code{\link[rlang:args_data_masking]{data-masked}}> Name of date
variable in \code{data}}
}
\value{
A length 1 numeric
}
\description{
Helper function that takes a data frame of data by sampling unit
and date, finds dates on which collar heights change, and uses those dates
to calculate a grand mean of collar height.
}
