% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gasmet_data.R
\name{read_gasmet_files}
\alias{read_gasmet_files}
\title{Read Gasmet TXT files as character}
\usage{
read_gasmet_files(paths = find_gasmet_files(), ...)
}
\arguments{
\item{paths}{Character vector. Paths of Gasmet TXT files. By default, the
paths are discovered using the default arguments of \code{\link[=find_gasmet_files]{find_gasmet_files()}}.}

\item{...}{Ignored.}
}
\value{
A list
}
\description{
Find and read Gasmet TXT files as character vectors for preliminary checks.
}
\seealso{
\code{\link[=find_gasmet_files]{find_gasmet_files()}}
}
