% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gasmet_data.R
\name{remove_empty_files}
\alias{remove_empty_files}
\title{Check Gasmet TXT file paths for empty files}
\usage{
remove_empty_files(paths, ...)
}
\arguments{
\item{paths}{A vector of Gasmet TXT file paths identified with
\link{find_gasmet_files}.}

\item{...}{Ignored.}
}
\value{
A vector of paths to empty files.
}
\description{
Check Gasmet TXT file paths for empty files
}
