% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gasmet_data.R
\name{remove_repeated_headers}
\alias{remove_repeated_headers}
\title{Remove repeated header rows from Gasmet TXT data}
\usage{
remove_repeated_headers(gasmet_files, ...)
}
\arguments{
\item{gasmet_files}{A list of Gasmet TXT data read using \link{read_gasmet_files}.}

\item{...}{Ignored.}
}
\value{
A list
}
\description{
Remove repeated header rows from Gasmet TXT data
}
