% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gasmet_data.R
\name{validate_gasmet_dates}
\alias{validate_gasmet_dates}
\title{Validate dates in data from Gasmet TXT files}
\usage{
validate_gasmet_dates(
  gasmet_data_list,
  path_n_date = -2,
  file_date_formats = c("\%m/\%d/\%Y", "\%Y-\%m-\%d"),
  path_date_formats = c("\%Y\%m\%d"),
  date_col = "Date"
)
}
\arguments{
\item{gasmet_data_list}{List of data frames from Gasmet TXT files, read using
\link{read_gasmet_files}.}

\item{path_n_date}{Integer. Location of date in path of Gasmet TXT file when
split by the file separator. Defaults to -2, which corresponds to the
directory containing the TXT file. Passed to \code{\link[stringr:str_split]{stringr::str_split_i()}}.}

\item{file_date_formats}{Character vector of possible date formats to parse
within TXT file data. Date parsing will be attempted in the order of
formats specified. See \link{strptime} for formatting abbreviations.}

\item{path_date_formats}{Character vector of possible formats to parse dates
extracted from file paths. Date parsing will be attempted in the order of
formats specified. See \link{strptime} for formatting abbreviations.}

\item{date_col}{Name of date column in Gasmet files. Defaults to "Date".}
}
\value{
A list with names "date_tests" (a data frame) and "message" (a
character vector summarizing 'bad' test results).
}
\description{
Check that dates in data from Gasmet TXT files (1) are available, (2) are
limited to one date per file, (3) match the dates indicated by the file path
containing the TXT files, and (4) are formatted in a recognizable way.
}
